% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-functions.R
\name{qz_init}
\alias{qz_init}
\alias{qz_connect}
\alias{list_printers}
\title{Interact with QZ from Server}
\usage{
qz_init(session)

qz_connect(session)

list_printers(session)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}}
}
\description{
Interact with QZ from Server
}
