% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-functions.R
\name{qz_launch}
\alias{qz_launch}
\alias{list_printers}
\alias{list_hid}
\alias{hid_read}
\title{Interact with QZ from Server}
\usage{
qz_launch(session)

list_printers(session)

list_hid(session)

hid_read(session, vendor, product)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}}

\item{vendor}{hex code for vendorId}

\item{product}{hex code for productId}
}
\description{
Interact with QZ from Server
}
