% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/qz-libs.R
\name{server_api}
\alias{server_api}
\alias{qz_libs}
\alias{rtray_lib}
\title{Include qz JavaScript libraries}
\usage{
server_api(session)

qz_libs()

rtray_lib()
}
\description{
Include qz JavaScript libraries
}
