% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signing.R
\name{qz_signature}
\alias{qz_signature}
\title{Sign messages from QZ Tray PEM file}
\usage{
qz_signature(toSign, myKey = Sys.getenv("qz_key"),
  password = Sys.getenv("qz_password"))
}
\arguments{
\item{toSign}{message to sign}
}
\description{
Sign messages from QZ Tray PEM file
}
