% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signing.R
\name{signature}
\alias{signature}
\title{Sign messages from QZ Tray PEM file}
\usage{
signature(toSign, myKey = Sys.getenv("qz_key"),
  password = Sys.getenv("qz_password"))
}
\arguments{
\item{toSign}{message to sign}

\item{myKey}{path to PEM key file}

\item{password}{password for PEM key file}
}
\description{
Sign messages from QZ Tray PEM file
}
